+++ src/ptwt/__init__.py
@@ -1,6 +1,8 @@
 """Differentiable and gpu enabled fast wavelet transforms in PyTorch."""
 from .continuous_transform import cwt
-from .conv_transform import wavedec, wavedec2, waverec, waverec2
+from .conv_transform import wavedec, waverec
+from .conv_transform_2 import wavedec2, waverec2
+from .conv_transform_3 import wavedec3, waverec3
 from .matmul_transform import MatrixWavedec, MatrixWaverec
 from .matmul_transform_2 import MatrixWavedec2, MatrixWaverec2
 from .packets import WaveletPacket, WaveletPacket2D
