+++ tests/test_boundary_filters.py
@@ -6,7 +6,7 @@ import pywt
 import scipy.signal
 import torch
 
-from src.ptwt.conv_transform import _flatten_2d_coeff_lst
+from ptwt.conv_transform import _flatten_2d_coeff_lst
 from src.ptwt.matmul_transform import (
     MatrixWavedec,
     MatrixWaverec,
